pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "admedi"
version = "0.1.0"
description = "Config-driven ad mediation management tool"
requires-python = ">=3.14"
license = "Apache-2.0"
dependencies = [
    "httpx>=0.27",
    "typer>=0.15",
    "fastmcp>=3.0.2",
    "pydantic>=2.0",
    "pyjwt>=2.8",
    "ruamel.yaml>=0.18",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
admedi = "admedi.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.15",
    "mypy>=1.13",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py314"

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests as integration tests (deselected by default)",
]
